[feature] - Utilizing the new Context API to avoid problem when legacy get removed
After reviewing react-redux and react-router refactoring to use the new Context API, one
of the major things that the authors did is to proive a usage for the new context API
but still supporting the older legacy APIs of React.

Using create-react-context as a polyfill will allow React users -16.3 and +16.3 to
use this fancy plugin without any problem regards future removal of the context API. This
plugin will use the new context API with fallback support for legacy context.